Update functional tests. Apps now standalone from tests. Use Bedrock.
