+++ tests/apps/bookstore/bookstore-mp/src/main/resources/META-INF/microprofile-config.properties
