+++ tests/apps/bookstore/bookstore-mp/src/main/resources/logging.properties
