+++ tests/apps/bookstore/bookstore-mp/src/test/resources/META-INF/microprofile-config.properties
