+++ tests/apps/bookstore/bookstore-mp/src/test/resources/book.json
