+++ tests/apps/bookstore/bookstore-se/src/main/resources/application.yaml
