+++ tests/apps/bookstore/bookstore-se/src/main/resources/logging.properties
