+++ tests/apps/bookstore/bookstore-se/src/test/resources/book.json
@@ -1,0 +1,17 @@
+{
+  "isbn": "123456",
+  "title": "The hunt for Red October",
+  "description": "387 p., hardback",
+  "summary": "The Soviets' new ballistic-missile submarine is attempting to defect to the United States, but the Soviet Atlantic fleet has been ordered to find and destroy her at all costs. Can Red October reach the U.S. safely?",
+  "genre": "spy stories",
+  "category": "fiction",
+  "publisher": "Naval Institute Press",
+  "copyright": "c1984",
+  "authors": [
+    {
+      "first": "Tom",
+      "last": "Clancy"
+    }
+  ]
+}
+
