+++ tests/apps/bookstore/common/src/main/resources/META-INF/beans.xml
