+++ tests/apps/bookstore/common/src/test/java/io/helidon/tests/apps/bookstore/mp/BookGeneratorTest.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2018, 2019 Oracle and/or its affiliates. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.tests.apps.bookstore.mp;
+
+import java.util.Collection;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+/**
+ * Class BookGeneratorTest.
+ */
+public class BookGeneratorTest {
+
+    @Test
+    public void testGenerator() {
+        Collection<Book> bookStore = new BookGenerator().apply(10);
+        assertEquals(10, bookStore.size());
+        bookStore.forEach(book -> System.out.println(book));
+    }
+}
