+++ tests/functional/bookstore/README.md
@@ -1,0 +1,11 @@
+Bookstore Functional Test
+
+This test runs and tests the Bookstore application. It 
+tests the following variations
+
+* SE Bookstore with JSON-P
+* SE Bookstore with JSON-B
+* SE Bookstore with Jackson
+* MP Bookstore with JSON-P (or whatever the application
+  is using
+
