+++ tests/functional/bookstore/src/test/resources/book.json
