--- tests/functional/helidon-mp/src/main/java/io/helidon/testing/mp/BookApplication.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2018, 2019 Oracle and/or its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.helidon.testing.mp;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-import java.util.Set;
-
-import io.helidon.common.CollectionsHelper;
-
-@ApplicationScoped
-@ApplicationPath("/")
-public class BookApplication extends Application {
-
-    @Override
-    public Set<Class<?>> getClasses() {
-        return CollectionsHelper.setOf(BookResource.class);
-    }
-}
